+++ src/client/shim.d.ts
@@ -1,4 +1,5 @@
 declare const __VP_HASH_MAP__: Record<string, string>
+declare const __VP_LOCAL_SEARCH__: boolean
 declare const __ALGOLIA__: boolean
 declare const __CARBON__: boolean
 declare const __VUE_PROD_DEVTOOLS__: boolean
@@ -24,4 +25,10 @@ declare module '@theme/index' {
 declare module '@localSearchIndex' {
   const data: Record<string, () => Promise<{ default: string }>>
   export default data
+}
+
+declare module 'mark.js/src/vanilla.js' {
+  import type { Mark } from 'mark.js'
+  const mark: Mark
+  export default mark
 }