+++ src/client/theme-default/components/VPLocalSearchBox.vue
@@ -1,13 +1,36 @@
 <script lang="ts" setup>
-import { markRaw, nextTick, onMounted, ref, shallowRef, watch, type Ref, createApp } from 'vue'
-import { useRouter } from 'vitepress'
-import { onKeyStroke, useSessionStorage, debouncedWatch, useLocalStorage, useEventListener, computedAsync } from '@vueuse/core'
-import MiniSearch, { type SearchResult } from 'minisearch'
 import localSearchIndex from '@localSearchIndex'
+import {
+  computedAsync,
+  debouncedWatch,
+  onKeyStroke,
+  useEventListener,
+  useLocalStorage,
+  useScrollLock,
+  useSessionStorage
+} from '@vueuse/core'
+import Mark from 'mark.js/src/vanilla.js'
+import MiniSearch, { type SearchResult } from 'minisearch'
+import { useRouter } from 'vitepress'
+import {
+  computed,
+  createApp,
+  markRaw,
+  nextTick,
+  onBeforeUnmount,
+  onMounted,
+  ref,
+  shallowRef,
+  watch,
+  watchEffect,
+  type Ref
+} from 'vue'
+import type { ModalTranslations } from '../../../../types/local-search'
+import { pathToFile } from '../../app/utils'
+import { slash } from '../../shared'
 import { useData } from '../composables/data'
 import { createTranslate } from '../support/translation'
-import type { ModalTranslations } from '../../../../types/local-search'
-import { pathToFile, withBase } from '../../app/utils.js'
+import { dataSymbol } from '../../app/data'
 
 defineProps<{
   placeholder: string
@@ -17,7 +40,9 @@ const emit = defineEmits<{
   (e: 'close'): void
 }>()
 
-const el = ref<HTMLDivElement>()
+const el = shallowRef<HTMLElement>()
+const resultsEl = shallowRef<HTMLElement>()
+const body = shallowRef<HTMLElement>()
 
 /* Search */
 
@@ -38,25 +63,55 @@ interface Result {
   text?: string
 }
 
-const { localeIndex } = useData()
+const vitePressData = useData()
+const { localeIndex, theme } = vitePressData
+const searchIndex = computedAsync(async () =>
+  markRaw(
+    MiniSearch.loadJSON<Result>(
+      (await searchIndexData.value[localeIndex.value]?.())?.default,
+      {
+        fields: ['title', 'titles', 'text'],
+        storeFields: ['title', 'titles'],
+        searchOptions: {
+          fuzzy: 0.2,
+          prefix: true,
+          boost: { title: 4, text: 2, titles: 1 }
+        }
+      }
+    )
+  )
+)
 
-const searchIndex = computedAsync(async () => markRaw(MiniSearch.loadJSON<Result>((await searchIndexData.value[localeIndex.value]?.())?.default, {
-  fields: ['title', 'titles', 'text'],
-  storeFields: ['title', 'titles'],
-  searchOptions: {
-    fuzzy: 0.2,
-    prefix: true,
-    boost: { title: 4, text: 2, titles: 1 },
-  },
-})))
+const disableQueryPersistence = computed(() => {
+  return (
+      theme.value.search?.provider === 'local' &&
+      theme.value.search.options?.disableQueryPersistence === true
+  )
+})
 
-const filterText = useSessionStorage('vitepress:local-search-filter', '')
+const filterText = disableQueryPersistence.value
+    ? ref('')
+    : useSessionStorage('vitepress:local-search-filter', '')
 
-const showDetailedList = useLocalStorage('vitepress:local-search-detailed-list', false)
+const showDetailedList = useLocalStorage(
+  'vitepress:local-search-detailed-list',
+  false
+)
+
+const disableDetailedView = computed(() => {
+  return (
+    theme.value.search?.provider === 'local' &&
+    theme.value.search.options?.disableDetailedView === true
+  )
+})
+
+watchEffect(() => {
+  if (disableDetailedView.value) {
+    showDetailedList.value = false
+  }
+})
 
 const results: Ref<(SearchResult & Result)[]> = shallowRef([])
-
-const contents = shallowRef(new Map<string, Map<string, string>>())
 
 const headingRegex = /<h(\d*).*?>.*?<a.*? href="#(.*?)".*?>.*?<\/a><\/h\1>/gi
 
@@ -66,93 +121,97 @@ watch(filterText, () => {
   enableNoResults.value = false
 })
 
-debouncedWatch(() => [searchIndex.value, filterText.value, showDetailedList.value] as const, async ([index, filterTextValue, showDetailedListValue], old, onCleanup) => {
-  let canceled = false
-  onCleanup(() => {
-    canceled = true
-  })
+const mark = computedAsync(async () => {
+  if (!resultsEl.value) return
+  return markRaw(new Mark(resultsEl.value))
+}, null)
 
-  if (!index) return
-
-  // Search
-  results.value = index.search(filterTextValue).slice(0, 16) as (SearchResult & Result)[]
-  enableNoResults.value = true
-
-  // Highlighting
-  const mods = showDetailedListValue ? await Promise.all(results.value.map(r => fetchExcerpt(r.id))) : []
-  if (canceled) return
-  const c = new Map<string, Map<string, string>>()
-  for (const { id, mod } of mods) {
-    const comp = mod.default ?? mod
-    if (comp?.render) {
-      const app = createApp(comp)
-      // Silence warnings about missing components
-      app.config.warnHandler = () => {}
-      const div = document.createElement('div')
-      app.mount(div)
-      const sections = div.innerHTML.split(headingRegex)
-      app.unmount()
-      sections.shift()
-      const mapId = id.slice(0, id.indexOf('#'))
-      let map = c.get(mapId)
-      if (!map) {
-        map = new Map()
-        c.set(mapId, map)
-      }
-      for (let i = 0; i < sections.length; i += 3) {
-        const anchor = sections[i + 1]
-        const html = sections[i + 2]
-        map.set(anchor, html)
-      }
-    }
-    if (canceled) return
-  }
-  results.value = results.value.map(r => {
-    let title = r.title
-    let titles = r.titles
-    let text = ''
-
-    // Highlight in text
-    const [id, anchor] = r.id.split('#')
-    const map = c.get(id)
-    if (map) {
-      text = map.get(anchor) ?? ''
-    }
-
-    for (const term in r.match) {
-      const match = r.match[term]
-      const reg = new RegExp(term, 'gi')
-      if (match.includes('title')) {
-        title = title.replace(reg, `<mark>$&</mark>`)
-      }
-      if (match.includes('titles')) {
-        titles = titles.map(t => t.replace(reg, `<mark>$&</mark>`))
-      }
-      if (showDetailedListValue && match.includes('text')) {
-        text = text.replace(reg, `<mark>$&</mark>`)
-      }
-    }
-
-    return {
-      ...r,
-      title,
-      titles,
-      text,
-    }
-  })
-  contents.value = c
-
-  await nextTick()
-  const excerpts = el.value?.querySelectorAll('.result .excerpt') ?? []
-  for (const excerpt of excerpts) {
-    excerpt.querySelector('mark')?.scrollIntoView({
-      block: 'center',
+debouncedWatch(
+  () => [searchIndex.value, filterText.value, showDetailedList.value] as const,
+  async ([index, filterTextValue, showDetailedListValue], old, onCleanup) => {
+    let canceled = false
+    onCleanup(() => {
+      canceled = true
     })
-  }
-}, { debounce: 200, immediate: true })
 
-async function fetchExcerpt (id: string) {
-  const file = pathToFile(withBase(id.slice(0, id.indexOf('#'))))
+    if (!index) return
+
+    // Search
+    results.value = index
+      .search(filterTextValue)
+      .slice(0, 16) as (SearchResult & Result)[]
+    enableNoResults.value = true
+
+    // Highlighting
+    const mods = showDetailedListValue
+      ? await Promise.all(results.value.map((r) => fetchExcerpt(r.id)))
+      : []
+    if (canceled) return
+    const c = new Map<string, Map<string, string>>()
+    for (const { id, mod } of mods) {
+      const comp = mod.default ?? mod
+      if (comp?.render) {
+        const app = createApp(comp)
+        // Silence warnings about missing components
+        app.config.warnHandler = () => {}
+        app.provide(dataSymbol, vitePressData)
+        const div = document.createElement('div')
+        app.mount(div)
+        const sections = div.innerHTML.split(headingRegex)
+        app.unmount()
+        sections.shift()
+        const mapId = id.slice(0, id.indexOf('#'))
+        let map = c.get(mapId)
+        if (!map) {
+          map = new Map()
+          c.set(mapId, map)
+        }
+        for (let i = 0; i < sections.length; i += 3) {
+          const anchor = sections[i + 1]
+          const html = sections[i + 2]
+          map.set(anchor, html)
+        }
+      }
+      if (canceled) return
+    }
+
+    const terms = new Set<string>()
+
+    results.value = results.value.map((r) => {
+      const [id, anchor] = r.id.split('#')
+      const map = c.get(id)
+      const text = map?.get(anchor) ?? ''
+      for (const term in r.match) {
+        terms.add(term)
+      }
+      return { ...r, text }
+    })
+
+    await nextTick()
+    if (canceled) return
+
+    await new Promise((r) => {
+      mark.value?.unmark({
+        done: () => {
+          mark.value?.markRegExp(formMarkRegex(terms), { done: r })
+        }
+      })
+    })
+
+    const excerpts = el.value?.querySelectorAll('.result .excerpt') ?? []
+    for (const excerpt of excerpts) {
+      excerpt
+        .querySelector('mark[data-markjs="true"]')
+        ?.scrollIntoView({ block: 'center' })
+    }
+    // FIXME: without this whole page scrolls to the bottom
+    resultsEl.value?.firstElementChild?.scrollIntoView({ block: 'start' })
+  },
+  { debounce: 200, immediate: true }
+)
+
+async function fetchExcerpt(id: string) {
+  const file = pathToFile(slash(id.slice(0, id.indexOf('#'))))
   try {
     return { id, mod: await import(/*@vite-ignore*/ file) }
   } catch (e) {
@@ -165,7 +224,7 @@ async function fetchExcerpt (id: string) {
 
 const searchInput = ref<HTMLInputElement>()
 
-function focusSearchInput () {
+function focusSearchInput() {
   searchInput.value?.focus()
   searchInput.value?.select()
 }
@@ -174,7 +233,7 @@ onMounted(() => {
   focusSearchInput()
 })
 
-function onSearchBarClick (event: PointerEvent) {
+function onSearchBarClick(event: PointerEvent) {
   if (event.pointerType === 'mouse') {
     focusSearchInput()
   }
@@ -190,12 +249,12 @@ watch(results, () => {
   scrollToSelectedResult()
 })
 
-function scrollToSelectedResult () {
+function scrollToSelectedResult() {
   nextTick(() => {
     const selectedEl = document.querySelector('.result.selected')
     if (selectedEl) {
       selectedEl.scrollIntoView({
-        block: 'nearest',
+        block: 'nearest'
       })
     }
   })
@@ -226,7 +285,7 @@ const router = useRouter()
 onKeyStroke('Enter', () => {
   const selectedPackage = results.value[selectedIndex.value]
   if (selectedPackage) {
-    router.go(withBase(selectedPackage.id))
+    router.go(selectedPackage.id)
     emit('close')
   }
 })
@@ -236,9 +295,6 @@ onKeyStroke('Escape', () => {
 })
 
 // Translations
-
-const { theme } = useData()
-
 const defaultTranslations: { modal: ModalTranslations } = {
   modal: {
     displayDetails: 'Display detailed list',
@@ -251,11 +307,13 @@ const defaultTranslations: { modal: ModalTranslations } = {
       navigateText: 'to navigate',
       navigateUpKeyAriaLabel: 'up arrow',
       navigateDownKeyAriaLabel: 'down arrow',
+      closeText: 'to close',
+      closeKeyAriaLabel: 'escape'
     }
   }
 }
 
-const $t = createTranslate(theme.value.localSearch, defaultTranslations)
+const $t = createTranslate(theme.value.search?.options, defaultTranslations)
 
 // Back
 
@@ -264,10 +322,38 @@ onMounted(() => {
   window.history.pushState(null, '', null)
 })
 
-useEventListener('popstate', event => {
+useEventListener('popstate', (event) => {
   event.preventDefault()
   emit('close')
 })
+
+/** Lock body */
+const isLocked = useScrollLock(body)
+
+onMounted(() => {
+  body.value = document.body
+  nextTick(() => {
+    isLocked.value = true
+  })
+})
+
+onBeforeUnmount(() => {
+  isLocked.value = false
+})
+
+function formMarkRegex(terms: Set<string>) {
+  return new RegExp(
+    [...terms]
+      .sort((a, b) => b.length - a.length)
+      .map((term) => {
+        return `(${term
+          .replace(/[|\\{}()[\]^$+*?.]/g, '\\$&')
+          .replace(/-/g, '\\x2d')})`
+      })
+      .join('|'),
+    'gi'
+  )
+}
 </script>
 
 <template>
@@ -277,14 +363,45 @@ useEventListener('popstate', event => {
 
       <div class="shell">
         <div class="search-bar" @pointerup="onSearchBarClick($event)">
-          <svg class="search-icon" width="16" height="16" viewBox="0 0 24 24"><g fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"><circle cx="11" cy="11" r="8"/><path d="m21 21l-4.35-4.35"/></g></svg>
+          <svg
+            class="search-icon"
+            width="18"
+            height="18"
+            viewBox="0 0 24 24"
+            aria-hidden="true"
+          >
+            <g
+              fill="none"
+              stroke="currentColor"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+              stroke-width="2"
+            >
+              <circle cx="11" cy="11" r="8" />
+              <path d="m21 21l-4.35-4.35" />
+            </g>
+          </svg>
           <div class="search-actions before">
             <button
               class="back-button"
               :title="$t('modal.backButtonTitle')"
               @click="$emit('close')"
             >
-              <svg width="16" height="16" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 12H5m7 7l-7-7l7-7"/></svg>
+              <svg
+                width="18"
+                height="18"
+                viewBox="0 0 24 24"
+                aria-hidden="true"
+              >
+                <path
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M19 12H5m7 7l-7-7l7-7"
+                />
+              </svg>
             </button>
           </div>
           <input
@@ -292,17 +409,30 @@ useEventListener('popstate', event => {
             v-model="filterText"
             :placeholder="placeholder"
             class="search-input"
-          >
+          />
           <div class="search-actions">
             <button
+              v-if="!disableDetailedView"
               class="toggle-layout-button"
-              :class="{
-                'detailed-list': showDetailedList,
-              }"
+              :class="{ 'detailed-list': showDetailedList }"
               :title="$t('modal.displayDetails')"
               @click="showDetailedList = !showDetailedList"
             >
-              <svg width="16" height="16" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 14h7v7H3zM3 3h7v7H3zm11 1h7m-7 5h7m-7 6h7m-7 5h7"/></svg>
+              <svg
+                width="18"
+                height="18"
+                viewBox="0 0 24 24"
+                aria-hidden="true"
+              >
+                <path
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M3 14h7v7H3zM3 3h7v7H3zm11 1h7m-7 5h7m-7 6h7m-7 5h7"
+                />
+              </svg>
             </button>
 
             <button
@@ -310,43 +440,63 @@ useEventListener('popstate', event => {
               :title="$t('modal.resetButtonTitle')"
               @click="filterText = ''"
             >
-              <svg width="16" height="16" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M20 5H9l-7 7l7 7h11a2 2 0 0 0 2-2V7a2 2 0 0 0-2-2Zm-2 4l-6 6m0-6l6 6"/></svg>
+              <svg
+                width="18"
+                height="18"
+                viewBox="0 0 24 24"
+                aria-hidden="true"
+              >
+                <path
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M20 5H9l-7 7l7 7h11a2 2 0 0 0 2-2V7a2 2 0 0 0-2-2Zm-2 4l-6 6m0-6l6 6"
+                />
+              </svg>
             </button>
           </div>
         </div>
 
         <div
+          ref="resultsEl"
           class="results"
           @mousemove="disableMouseOver = false"
         >
           <a
             v-for="(p, index) in results"
             :key="p.id"
-            :href="withBase(p.id)"
+            :href="p.id"
             class="result"
             :class="{
-              selected: selectedIndex === index,
+              selected: selectedIndex === index
             }"
-            :aria-title="[...p.titles, p.title].join(' > ')"
+            :aria-label="[...p.titles, p.title].join(' > ')"
             @mouseenter="!disableMouseOver && (selectedIndex = index)"
             @click="$emit('close')"
           >
             <div>
               <div class="titles">
                 <span class="title-icon">#</span>
-                <span
-                  v-for="(t, index) in p.titles"
-                  :key="index"
-                  class="title"
-                >
+                <span v-for="(t, index) in p.titles" :key="index" class="title">
                   <span class="text" v-html="t" />
-                  <svg width="16" height="16" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="m9 18l6-6l-6-6"/></svg>
+                  <svg width="18" height="18" viewBox="0 0 24 24">
+                    <path
+                      fill="none"
+                      stroke="currentColor"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="m9 18l6-6l-6-6"
+                    />
+                  </svg>
                 </span>
                 <span class="title main">
                   <span class="text" v-html="p.title" />
                 </span>
               </div>
-              
+
               <div v-if="showDetailedList" class="excerpt-wrapper">
                 <div v-if="p.text" class="excerpt">
                   <div class="vp-doc" v-html="p.text" />
@@ -357,19 +507,63 @@ useEventListener('popstate', event => {
             </div>
           </a>
 
-          <div v-if="filterText && !results.length && enableNoResults" class="no-results">
-            {{ $t('modal.noResultsText') }} "<strong>{{ filterText }}</strong>"
+          <div
+            v-if="filterText && !results.length && enableNoResults"
+            class="no-results"
+          >
+            {{ $t('modal.noResultsText') }} "<strong>{{ filterText }}</strong
+            >"
           </div>
         </div>
 
         <div class="search-keyboard-shortcuts">
           <span>
-            <kbd :aria-title="$t('modal.footer.navigateUpKeyAriaLabel')"><svg width="14" height="14" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 19V5m-7 7l7-7l7 7"/></svg></kbd> <kbd :aria-title="$t('modal.footer.navigateDownKeyAriaLabel')"><svg width="14" height="14" viewBox="0 0 24 24"><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 5v14m7-7l-7 7l-7-7"/></svg></kbd>
+            <kbd :aria-label="$t('modal.footer.navigateUpKeyAriaLabel')">
+              <svg width="14" height="14" viewBox="0 0 24 24">
+                <path
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M12 19V5m-7 7l7-7l7 7"
+                />
+              </svg>
+            </kbd>
+            <kbd :aria-label="$t('modal.footer.navigateDownKeyAriaLabel')">
+              <svg width="14" height="14" viewBox="0 0 24 24">
+                <path
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                  d="M12 5v14m7-7l-7 7l-7-7"
+                />
+              </svg>
+            </kbd>
             {{ $t('modal.footer.navigateText') }}
           </span>
           <span>
-            <kbd :aria-title="$t('modal.footer.selectKeyAriaLabel')"><svg width="14" height="14" viewBox="0 0 24 24"><g fill="none" stroke="currentcolor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"><path d="m9 10l-5 5l5 5"/><path d="M20 4v7a4 4 0 0 1-4 4H4"/></g></svg></kbd>
+            <kbd :aria-label="$t('modal.footer.selectKeyAriaLabel')">
+              <svg width="14" height="14" viewBox="0 0 24 24">
+                <g
+                  fill="none"
+                  stroke="currentcolor"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  stroke-width="2"
+                >
+                  <path d="m9 10l-5 5l5 5" />
+                  <path d="M20 4v7a4 4 0 0 1-4 4H4" />
+                </g>
+              </svg>
+            </kbd>
             {{ $t('modal.footer.selectText') }}
+          </span>
+          <span>
+            <kbd :aria-label="$t('modal.footer.closeKeyAriaLabel')">esc</kbd>
+            {{ $t('modal.footer.closeText') }}
           </span>
         </div>
       </div>
@@ -377,7 +571,7 @@ useEventListener('popstate', event => {
   </Teleport>
 </template>
 
-<style scoped lang="postcss">
+<style scoped>
 .VPLocalSearchBox {
   position: fixed;
   z-index: 100;
@@ -388,7 +582,8 @@ useEventListener('popstate', event => {
 .backdrop {
   position: absolute;
   inset: 0;
-  background: rgba(0, 0, 0, 0.5);
+  background: var(--vp-backdrop-bg-color);
+  transition: opacity 0.5s;
 }
 
 .shell {
@@ -398,7 +593,7 @@ useEventListener('popstate', event => {
   display: flex;
   flex-direction: column;
   gap: 16px;
-  background: var(--vp-c-bg);
+  background: var(--vp-local-search-bg);
   width: min(100vw - 60px, 900px);
   height: min-content;
   max-height: min(100vh - 128px, 900px);
@@ -416,7 +611,7 @@ useEventListener('popstate', event => {
 }
 
 .search-bar {
-  border: 1px solid rgba(128, 128, 128, 0.2);
+  border: 1px solid var(--vp-c-divider);
   border-radius: 4px;
   display: flex;
   align-items: center;
@@ -432,6 +627,10 @@ useEventListener('popstate', event => {
 
 .search-bar:focus-within {
   border-color: var(--vp-c-brand);
+}
+
+.search-icon {
+  margin: 8px;
 }
 
 @media (max-width: 768px) {
@@ -454,6 +653,13 @@ useEventListener('popstate', event => {
 
 .search-actions {
   display: flex;
+  gap: 4px;
+}
+
+@media (any-pointer: coarse) {
+  .search-actions {
+    gap: 8px;
+  }
 }
 
 @media (min-width: 769px) {
@@ -463,7 +669,7 @@ useEventListener('popstate', event => {
 }
 
 .search-actions button {
-  padding: 8px 6px;
+  padding: 8px;
 }
 
 .search-actions button:hover,
@@ -477,6 +683,13 @@ useEventListener('popstate', event => {
   display: flex;
   flex-wrap: wrap;
   gap: 16px;
+  line-height: 14px;
+}
+
+.search-keyboard-shortcuts span {
+  display: flex;
+  align-items: center;
+  gap: 4px;
 }
 
 @media (max-width: 768px) {
@@ -503,6 +716,7 @@ useEventListener('popstate', event => {
   gap: 6px;
   overflow-x: hidden;
   overflow-y: auto;
+  overscroll-behavior: contain;
 }
 
 .result {
@@ -512,7 +726,7 @@ useEventListener('popstate', event => {
   border-radius: 4px;
   transition: none;
   line-height: 1rem;
-  border: solid 2px rgba(128, 128, 128, 0.05);
+  border: solid 2px var(--vp-local-search-result-border);
 }
 
 .result > div {
@@ -557,7 +771,8 @@ useEventListener('popstate', event => {
 }
 
 .result.selected {
-  border-color: var(--vp-c-brand);
+  --vp-local-search-result-bg: var(--vp-local-search-result-selected-bg);
+  border-color: var(--vp-local-search-result-selected-border);
 }
 
 .excerpt-wrapper {
@@ -585,9 +800,10 @@ useEventListener('popstate', event => {
 
 .titles :deep(mark),
 .excerpt :deep(mark) {
-  background-color: var(--vp-c-highlight-bg);
-  color: var(--vp-c-highlight-text);
+  background-color: var(--vp-local-search-highlight-bg);
+  color: var(--vp-local-search-highlight-text);
   border-radius: 2px;
+  padding: 0 2px;
 }
 
 .excerpt :deep(.vp-code-group) .tabs {
@@ -604,7 +820,7 @@ useEventListener('popstate', event => {
   left: 0;
   width: 100%;
   height: 8px;
-  background: linear-gradient(transparent, var(--vp-c-bg));
+  background: linear-gradient(transparent, var(--vp-local-search-result-bg));
   z-index: 1000;
 }
 
@@ -614,7 +830,7 @@ useEventListener('popstate', event => {
   left: 0;
   width: 100%;
   height: 8px;
-  background: linear-gradient(var(--vp-c-bg), transparent);
+  background: linear-gradient(var(--vp-local-search-result-bg), transparent);
   z-index: 1000;
 }